Rename extensions-packs.ts to extension-pack-picker.ts
